pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "bug-triage-env"
version = "1.0.0"
description = "Bug Triage & Fix Recommendation — OpenEnv Environment"
requires-python = ">=3.10"
dependencies = [
    "openenv-core",
    "fastapi",
    "uvicorn[standard]",
    "pydantic>=2.0",
    "openai",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "httpx",
]

[project.scripts]
server = "server.app:main"

[tool.setuptools.packages.find]
include = ["bug_triage_env*", "server*"]
